keras_server/requirements.txt
tensorflow-macos
tensorflow-metal
fastapi
uvicorn[standard]
numpy
